refactorisation pour injecter les sessions dans les routes
